refactor: fix imports, remove refs to db_models
